refactor(split-reveal): composable API with modular subfiles
Opt-in Images and Task primitives, external ready/progress control, and
tree-shakeable modules. Registry bundler follows co-located imports; photographer
demo uses the default Progress UI.
