+++ animata/preloader/split-reveal/execute-task.ts
@@ -1,0 +1,29 @@
+import type { SplitRevealProgressReporter, SplitRevealTaskDefinition } from "./types";
+
+export async function executeTask(
+  task: SplitRevealTaskDefinition,
+  report: SplitRevealProgressReporter,
+  signal: AbortSignal,
+) {
+  const ctx = { report, signal };
+
+  if (task.run) {
+    await task.run(ctx);
+    return;
+  }
+
+  if (task.generator) {
+    const iterator = task.generator(ctx);
+    let result = await iterator.next();
+
+    while (!result.done) {
+      if (signal.aborted) {
+        return;
+      }
+      if (result.value) {
+        report(result.value);
+      }
+      result = await iterator.next();
+    }
+  }
+}
