+++ content/docs/changelog/2026-06.mdx
@@ -95,6 +95,10 @@ Ran a site-wide react-doctor remediation pass and cleared the scan to zero error
     <h3>Flipping Cards</h3>
     <p>Replaced the `list` prop and inline flip markup with `FlippingCards`, `Item`, `Item.Front`, and `Item.Back` — grid layout on the list, flip mechanics delegated to Flip Card.</p>
   </ChangeLogEntry>
+  <ChangeLogEntry href="/docs/preloader/split-reveal">
+    <h3>Split Reveal</h3>
+    <p>Moved image preloading to opt-in `SplitReveal.Images`. Added `SplitReveal.Task` for custom async or generator progress, plus `ready` and `progress` props for external boot control. Shutters and progress UI compose under `SplitReveal.Overlay`.</p>
+  </ChangeLogEntry>
 </ChangeLogComponents>
 
 ## Documentation
