+++ content/docs/preloader/index.mdx
@@ -5,3 +5,9 @@ author: sudhashrestha
 ---
 
 <CategoryIndex category="preloader" />
+
+## Recent changes
+
+| Date | Component | Change |
+| --- | --- | --- |
+| 2026-06 | [Split Reveal](/docs/preloader/split-reveal) | Composable `Images`, `Task`, `Overlay`; modular source under `split-reveal/` |
